move rectangle grid rotation handler to its own class
